# openapi.yaml
openapi: 3.0.0
info:
  description: API for managing posts and categories on a website
  title: Project Peak API
  version: 1.1.0
servers:
- description: Production server
  url: https://api.example.com/api/v1
- description: Development server
  url: http://localhost:8000/api/v1
...
paths:
...
  /categories:
    get:
      responses:
        "200":
          content:
            application/json:
              examples:
                categoriesList:
                  summary: List of categories retrieved successfully
                  value:
                  - id: c1bfb8a0-3b84-4f43-949e-3a7f4f2a1e21
                    name: Fitness
                    description: Posts related to fitness and exercise routines.
                  - id: 9a4f8d6c-03c5-4f4f-b291-6a1b8bcb7f3e
                    name: Nutrition
                    description: Posts related to healthy eating and nutrition.
              schema:
                items:
                  $ref: '#/components/schemas/Category'
                type: array
          description: List of categories retrieved successfully
        "401":
          content:
            application/json:
              examples:
                example:
                  summary: Unauthorized error example
                  value:
                    statusCode: 401
                    message: Authentication required to access this resource
                    error: Unauthorized
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized. The client must authenticate itself to get the
            requested response.
        "403":
          content:
            application/json:
              examples:
                example:
                  summary: Forbidden error example
                  value:
                    statusCode: 403
                    message: You don't have permission to access this resource
                    error: Forbidden
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden. The client does not have access rights to the content.
        "500":
          content:
            application/json:
              examples:
                example:
                  summary: Internal server error example
                  value:
                    statusCode: 500
                    message: An unexpected error occurred on the server
                    error: Internal Server Error
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error. The server has encountered a situation
            it doesn't know how to handle.
      summary: Get all categories
      tags:
      - Category
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
        required: true
      responses:
        "201":
          content:
            application/json:
              examples:
                categoryCreated:
                  summary: Category created successfully
                  value:
                    id: c1bfb8a0-3b84-4f43-949e-3a7f4f2a1e21
                    name: Fitness
                    description: Posts related to fitness and exercise routines.
              schema:
                $ref: '#/components/schemas/Category'
          description: Category created successfully
        "401":
          content:
            application/json:
              examples:
                example:
                  summary: Unauthorized error example
                  value:
                    statusCode: 401
                    message: Authentication required to access this resource
                    error: Unauthorized
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized. The client must authenticate itself to get the
            requested response.
        "403":
          content:
            application/json:
              examples:
                example:
                  summary: Forbidden error example
                  value:
                    statusCode: 403
                    message: You don't have permission to access this resource
                    error: Forbidden
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden. The client does not have access rights to the content.
        "500":
          content:
            application/json:
              examples:
                example:
                  summary: Internal server error example
                  value:
                    statusCode: 500
                    message: An unexpected error occurred on the server
                    error: Internal Server Error
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error. The server has encountered a situation
            it doesn't know how to handle.
      summary: Create a new category
      tags:
      - Category
  /categories/{categoryId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: categoryId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Category deleted
        "401":
          content:
            application/json:
              examples:
                example:
                  summary: Unauthorized error example
                  value:
                    statusCode: 401
                    message: Authentication required to access this resource
                    error: Unauthorized
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized. The client must authenticate itself to get the
            requested response.
        "403":
          content:
            application/json:
              examples:
                example:
                  summary: Forbidden error example
                  value:
                    statusCode: 403
                    message: You don't have permission to access this resource
                    error: Forbidden
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden. The client does not have access rights to the content.
        "404":
          content:
            application/json:
              examples:
                example:
                  summary: Not found error example
                  value:
                    statusCode: 404
                    message: The requested resource was not found
                    error: Not Found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found. The server can not find the requested resource.
        "500":
          content:
            application/json:
              examples:
                example:
                  summary: Internal server error example
                  value:
                    statusCode: 500
                    message: An unexpected error occurred on the server
                    error: Internal Server Error
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error. The server has encountered a situation
            it doesn't know how to handle.
      summary: Delete a category
      tags:
      - Category
    get:
      parameters:
      - explode: false
        in: path
        name: categoryId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                categoryRetrieved:
                  summary: Category retrieved successfully
                  value:
                    id: c1bfb8a0-3b84-4f43-949e-3a7f4f2a1e21
                    name: Fitness
                    description: Posts related to fitness and exercise routines.
              schema:
                $ref: '#/components/schemas/Category'
          description: Category retrieved successfully
        "401":
          content:
            application/json:
              examples:
                example:
                  summary: Unauthorized error example
                  value:
                    statusCode: 401
                    message: Authentication required to access this resource
                    error: Unauthorized
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized. The client must authenticate itself to get the
            requested response.
        "403":
          content:
            application/json:
              examples:
                example:
                  summary: Forbidden error example
                  value:
                    statusCode: 403
                    message: You don't have permission to access this resource
                    error: Forbidden
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden. The client does not have access rights to the content.
        "404":
          content:
            application/json:
              examples:
                example:
                  summary: Not found error example
                  value:
                    statusCode: 404
                    message: The requested resource was not found
                    error: Not Found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found. The server can not find the requested resource.
        "500":
          content:
            application/json:
              examples:
                example:
                  summary: Internal server error example
                  value:
                    statusCode: 500
                    message: An unexpected error occurred on the server
                    error: Internal Server Error
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error. The server has encountered a situation
            it doesn't know how to handle.
      summary: Get a single category
      tags:
      - Category
    put:
      parameters:
      - explode: false
        in: path
        name: categoryId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                categoryUpdated:
                  summary: Category updated successfully
                  value:
                    id: c1bfb8a0-3b84-4f43-949e-3a7f4f2a1e21
                    name: Fitness
                    description: Posts related to fitness and exercise routines.
              schema:
                $ref: '#/components/schemas/Category'
          description: Category updated successfully
        "401":
          content:
            application/json:
              examples:
                example:
                  summary: Unauthorized error example
                  value:
                    statusCode: 401
                    message: Authentication required to access this resource
                    error: Unauthorized
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized. The client must authenticate itself to get the
            requested response.
        "403":
          content:
            application/json:
              examples:
                example:
                  summary: Forbidden error example
                  value:
                    statusCode: 403
                    message: You don't have permission to access this resource
                    error: Forbidden
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden. The client does not have access rights to the content.
        "404":
          content:
            application/json:
              examples:
                example:
                  summary: Not found error example
                  value:
                    statusCode: 404
                    message: The requested resource was not found
                    error: Not Found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found. The server can not find the requested resource.
        "500":
          content:
            application/json:
              examples:
                example:
                  summary: Internal server error example
                  value:
                    statusCode: 500
                    message: An unexpected error occurred on the server
                    error: Internal Server Error
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error. The server has encountered a situation
            it doesn't know how to handle.
      summary: Update a category
      tags:
      - Category
components:
...
  schemas:
...
    Category:
      example:
        name: Fitness
        description: Posts related to fitness and exercise routines.
        id: c1bfb8a0-3b84-4f43-949e-3a7f4f2a1e21
      properties:
        id:
          description: Unique identifier for the category
          example: c1bfb8a0-3b84-4f43-949e-3a7f4f2a1e21
          title: id
          type: string
        name:
          description: Name of the category
          example: Fitness
          title: name
          type: string
        description:
          description: A brief description of the category
          example: Posts related to fitness and exercise routines.
          title: description
          type: string
      title: Category
      type: object
    CategoryCreate:
      allOf:
      - $ref: '#/components/schemas/Category'
      example:
        name: Fitness
        description: Posts related to fitness and exercise routines.
        id: c1bfb8a0-3b84-4f43-949e-3a7f4f2a1e21
      required:
      - description
      - name
      title: CategoryCreate
    CategoryUpdate:
      allOf:
      - $ref: '#/components/schemas/Category'
      example:
        name: Fitness
        description: Posts related to fitness and exercise routines.
        id: c1bfb8a0-3b84-4f43-949e-3a7f4f2a1e21
      required:
      - description
      - name
      title: CategoryUpdate
    ErrorResponse:
      example:
        message: message
        error: error
        statusCode: 5
      properties:
        statusCode:
          description: HTTP status code
          title: statusCode
          type: integer
        message:
          description: Error message
          title: message
          type: string
        error:
          description: Error type
          title: error
          type: string
      title: ErrorResponse
      type: object
...